# perf(airbyte-cdk): performance enhancement:: add PrintBuffer