--- a/airbyte-cdk/python/airbyte_cdk/utils/__init__.py
+++ b/airbyte-cdk/python/airbyte_cdk/utils/__init__.py
@@ -5,5 +5,6 @@
 from .is_cloud_environment import is_cloud_environment
 from .schema_inferrer import SchemaInferrer
 from .traced_exception import AirbyteTracedException
+from .print_buffer import PrintBuffer
 
-__all__ = ["AirbyteTracedException", "SchemaInferrer", "is_cloud_environment"]
+__all__ = ["AirbyteTracedException", "SchemaInferrer", "is_cloud_environment", "PrintBuffer"]