--- a/airbyte-cdk/python/airbyte_cdk/utils/print_buffer.py
+++ b/airbyte-cdk/python/airbyte_cdk/utils/print_buffer.py
@@ -4,16 +4,18 @@ import time
 from io import StringIO
 import sys
 from threading import RLock
+from types import TracebackType
+from typing import Optional
 
 
 class PrintBuffer:
-    def __init__(self, flush_interval=0.1):
+    def __init__(self, flush_interval: float = 0.1):
         self.buffer = StringIO()
         self.flush_interval = flush_interval
         self.last_flush_time = time.monotonic()
         self.lock = RLock()
 
-    def write(self, message):
+    def write(self, message: str) -> None:
         with self.lock:
             self.buffer.write(message)
             current_time = time.monotonic()
@@ -21,16 +23,16 @@ class PrintBuffer:
                 self.flush()
                 self.last_flush_time = current_time
 
-    def flush(self):
+    def flush(self) -> None:
         with self.lock:
             combined_message = self.buffer.getvalue()
-            sys.__stdout__.write(combined_message)
+            sys.__stdout__.write(combined_message)  # type: ignore[union-attr]
             self.buffer = StringIO()
 
-    def __enter__(self):
+    def __enter__(self) -> "PrintBuffer":
         self.old_stdout = sys.stdout
         sys.stdout = self
         return self
 
-    def __exit__(self, exc_type, exc_val, exc_tb):
+    def __exit__(self, exc_type: Optional[BaseException], exc_val: Optional[BaseException], exc_tb: Optional[TracebackType]) -> None:
         self.flush()