## What

- Resolve https://github.com/airbytehq/airbyte-internal-issues/issues/8860
- speed up output


## How

use Buffer to catch stdout and flush every 0.1 sec

>[!NOTE]
`RLock()` is used to support multuthreading, because StringIO is not thread-safe.

## Review guide

1. `airbyte-cdk/python/airbyte_cdk/entrypoint.py`
2. `airbyte-cdk/python/airbyte_cdk/utils/print_buffer.py`


## User Impact

No

### Test in Docker; 2_000_000 records; stream: dummy_records
`/usr/bin/time -h docker run --rm -v $(pwd)/secrets:/secrets -v $(pwd)/integration_tests:/integration_tests airbyte/source-hardcoded-records:dev read --config /secrets/config.json --catalog /integration_tests/configured_catalog.json > /tmp/test.txt`

| TIME | real | user | sys |
|--------|--------|--------|--------|
| Before | 1m11.74s | 3.83s | 11.74s |
| After | 45.81s (1.56x) | 0.73s| 2.75s | 


## Can this PR be safely reverted and rolled back?
<!--
* If unsure, leave it blank.
-->
- [ ] YES 💚
- [ ] NO ❌