@artem1205 2024-07-26
/approve-regression-tests
RUN (successful) : https://github.com/airbytehq/airbyte/actions/runs/10116506269/job/27979488515